Always ask for the full 32-bytes sense buffer.
We allocate the buffer ourselves, so there's no reason to run `len()` on it
to get the length. And as it turns out, since we're using a pointer rather
than a view, this is using the _string_ length rather than the buffer
length.

This should correct issue #2. Tested using
python-scsi/examples/reportluns.py on a device that does not have LUNs.