Allow callers to request a different amount of sense data.
Since sense data is not restricted to 32 bytes (but 32 bytes cover most of
the common usage), it's possible that the callers would want to receive
more sense data.

Keep the logic to allocate the sense data in Cython and returning however
much was returned to Python in the CheckConditionError exception.